cheri: Resolve compiler warnings in lib/ and kernel/
Resolve remaining compiler warnings in lib/ and kernel/.

+++ include/linux/maple_tree.h
@@ -464,7 +464,7 @@ struct ma_wr_state {
  * top of the tree as the tree may have been modified.
  */
 #define MA_ERROR(err) \
-		((struct maple_enode *)(((unsigned long)err << 2) | 2UL))
+		((struct maple_enode *)(((uintptr_t)err << 2) | 2UL))
 
 #define MA_STATE(name, mt, first, end)					\
 	struct ma_state name = {					\
